messaging now perfect, you can simultanously see the lobbies info
